NER-83 IMP: changed typicalweight datatype to bigdecimal

+++ src/main/java/models/StockItem.java
@@ -1,6 +1,8 @@
 package models;
 
 import jakarta.persistence.*;
+
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
@@ -22,7 +24,7 @@ public class StockItem {
     private double RecommendedRetailPrice;
 
     @Column(nullable = false)
-    private double TypicalWeightPerUnit;
+    private BigDecimal TypicalWeightPerUnit;
 
     @Column(length = 11, nullable = false)
     private int QuantityPerOuter;
@@ -42,7 +44,7 @@ public class StockItem {
         return stockItemName;
     }
 
-    public double getTypicalWeightPerUnit() {
+    public BigDecimal getTypicalWeightPerUnit() {
         return TypicalWeightPerUnit;
     }
 }
